computer/requirements.txt
# Core dependencies
numpy>=1.24
opencv-python>=4.8

# Bluetooth — two transport back-ends are provided:
#
#   1. RFCOMMTransport  (Linux preferred)
#      Uses raw AF_BLUETOOTH/BTPROTO_RFCOMM sockets — no extra library needed.
#
#   2. SerialTransport  (macOS preferred)
#      After pairing the ESP32 in System Settings → Bluetooth, macOS creates a
#      /dev/cu.RobotESP32-xxx device. Use SerialTransport with that path.
pyserial>=3.5

# Optional: for scanning / discovering the ESP32 MAC address on Linux
# pybluez>=0.23
